requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
pydantic==1.10.4
PyJWT==2.6.0
python-dotenv==0.21.0
python-telegram-bot==20.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sarufi==0.1.2
sniffio==1.3.0
starlette==0.22.0
twilio==7.16.0
typing_extensions==4.4.0
unicorn==2.0.1.post1
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
